+++ pipestat/backends/db_backend/db_parsed_schema.py
@@ -5,7 +5,9 @@ import datetime
 import logging
 from pathlib import Path
 from typing import Any, Dict, List, Mapping, Optional
-from pydantic import create_model
+
+from pydantic import ConfigDict, create_model
+
 
 from sqlalchemy import Column, null
 from sqlalchemy.dialects.postgresql import JSONB
@@ -55,8 +57,7 @@ def get_base_model():
     class BaseModel(SQLModel):
         __table_args__ = {"extend_existing": True}
 
-        class Config:
-            arbitrary_types_allowed = True
+        model_config = ConfigDict(arbitrary_types_allowed=True)
 
     # return SQLModel
     return BaseModel
@@ -172,7 +173,7 @@ class ParsedSchemaDB(ParsedSchema):
                     # Optional[str],
                     # CLASSES_BY_TYPE[subdata[SCHEMA_TYPE_KEY]],
                     data_type,
-                    Field(default=subdata.get("default")),
+                    Field(default=subdata.get("default"), nullable=True),
                 )
         return defs
 
@@ -232,7 +233,8 @@ class ParsedSchemaDB(ParsedSchema):
             raise SchemaError(
                 f"'{PROJECT_NAME}' is reserved as identifier and can't be part of schema."
             )
-        field_defs[PROJECT_NAME] = (str, Field(default=None))
+
+        field_defs[PROJECT_NAME] = (str, Field(default=None, nullable=True))
 
         return field_defs
 
@@ -242,7 +244,8 @@ class ParsedSchemaDB(ParsedSchema):
             raise SchemaError(
                 f"'{PIPELINE_NAME}' is reserved as identifier and can't be part of schema."
             )
-        field_defs[PIPELINE_NAME] = (str, Field(default=None))
+
+        field_defs[PIPELINE_NAME] = (str, Field(default=None, nullable=True))
 
         return field_defs
 
@@ -264,7 +267,9 @@ class ParsedSchemaDB(ParsedSchema):
             raise SchemaError(
                 f"'{RECORD_IDENTIFIER}' is reserved as identifier and can't be part of schema."
             )
-        field_defs[RECORD_IDENTIFIER] = (str, Field(default=None))
+
+        field_defs[RECORD_IDENTIFIER] = (str, Field(default=None, nullable=True))
+
         return field_defs
 
     @staticmethod
@@ -273,7 +278,9 @@ class ParsedSchemaDB(ParsedSchema):
             raise SchemaError(
                 f"'{SAMPLE_NAME}' is reserved as identifier and can't be part of schema."
             )
-        field_defs[SAMPLE_NAME] = (str, Field(default=None))
+
+        field_defs[SAMPLE_NAME] = (str, Field(default=None, nullable=True))
+
         return field_defs
 
     @staticmethod
@@ -282,7 +289,9 @@ class ParsedSchemaDB(ParsedSchema):
             raise SchemaError(
                 f"'{STATUS}' is reserved for status reporting and can't be part of schema."
             )
-        field_defs[STATUS] = (str, Field(default=None))
+
+        field_defs[STATUS] = (str, Field(default=None, nullable=True))
+
         return field_defs
 
     @staticmethod
@@ -291,7 +300,9 @@ class ParsedSchemaDB(ParsedSchema):
             raise SchemaError(
                 f"'{CREATED_TIME}' is reserved for time reporting and can't be part of schema."
             )
-        field_defs[CREATED_TIME] = (datetime.datetime, Field(default=None))
+
+        field_defs[CREATED_TIME] = (datetime.datetime, Field(default=None, nullable=True))
+
         return field_defs
 
     @staticmethod
@@ -300,7 +311,9 @@ class ParsedSchemaDB(ParsedSchema):
             raise SchemaError(
                 f"'{MODIFIED_TIME}' is reserved for time reporting and can't be part of schema."
             )
-        field_defs[MODIFIED_TIME] = (datetime.datetime, Field(default=None))
+
+        field_defs[MODIFIED_TIME] = (datetime.datetime, Field(default=None, nullable=True))
+
         return field_defs
 
     def _table_name(self, suffix: str) -> str: