+++ docs/changelog.md
@@ -3,6 +3,14 @@
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
 
+## [0.7.0] - 2024-01-17
+### Added
+- `__iter__` now takes limit and cursor arguments to create an iterator from `select_records`
+### Changed
+- updated pydantic requirement to be >= 2.5.3
+### Fixed
+- Get summary files: Objects YAML button now works.
+
 ## [0.6.0] - 2023-12-22
 ### Added
 - `select_records`, which allows for a single API for selecting attributes (result_identifiers) given filter_conditions and/or columns